EP-4019: offload polling SHub batch processes to async task

+++ openeogeotrellis/async_task.py
@@ -1,6 +1,9 @@
 import argparse
 import json
 import logging
+import os
+import sys
+import time
 from typing import List
 
 import openeogeotrellis
@@ -9,27 +12,47 @@ from openeogeotrellis.backend import GpsBatchJobs
 from openeogeotrellis.configparams import ConfigParams
 from py4j.java_gateway import JavaGateway
 
+from openeogeotrellis.job_registry import JobRegistry
+
 logging.basicConfig(level=logging.INFO)
 openeogeotrellis.backend.logger.setLevel(logging.DEBUG)
 
 _log = logging.getLogger(__name__)
 
 ASYNC_TASK_ENDPOINT = "http://127.0.0.1:7180/asynctask"
+SENTINEL_HUB_BATCH_PROCESSES_POLL_INTERVAL_S = 60
+
+TASK_DELETE_BATCH_PROCESS_RESULTS = 'delete_batch_process_results'
+TASK_POLL_SENTINELHUB_BATCH_PROCESSES = 'poll_sentinelhub_batch_processes'
 
 
 def schedule_delete_batch_process_results(batch_job_id: str, subfolders: List[str]):
-    _schedule_task(
-        task_id="delete_batch_process_results",
-        arguments={
-            "batch_job_id": batch_job_id,
-            "subfolders": subfolders
-        })
+    _schedule_task(task_id=TASK_DELETE_BATCH_PROCESS_RESULTS,
+                   arguments={
+                       'batch_job_id': batch_job_id,
+                       'subfolders': subfolders
+                   })
 
 
-def _schedule_task(task_id: str, arguments: dict):
+def schedule_poll_sentinelhub_batch_processes(batch_job_id: str, user_id: str):
+    _schedule_task(task_id=TASK_POLL_SENTINELHUB_BATCH_PROCESSES,
+                   arguments={
+                       'batch_job_id': batch_job_id,
+                       'user_id': user_id
+                   },
+                   environment={
+                       'BATCH_JOBS_ZOOKEEPER_ROOT_PATH': ConfigParams().batch_jobs_zookeeper_root_path
+                   })
+
+
+def _schedule_task(task_id: str, arguments: dict, environment=None):
+    if environment is None:
+        environment = {}
+
     resp = requests.post(url=ASYNC_TASK_ENDPOINT, json={
-        "task_id": task_id,
-        "arguments": arguments
+        'task_id': task_id,
+        'arguments': arguments,
+        'environment': environment
     })
 
     resp.raise_for_status()
@@ -37,7 +60,7 @@ def _schedule_task(task_id: str, arguments: dict):
 
 # TODO: DRY this, cleaner.sh and job_tracker.sh
 def main():
-    _log.info("ConfigParams(): {c}".format(c=ConfigParams()))
+    _log.info("argv: {a!r}".format(a=sys.argv))
 
     parser = argparse.ArgumentParser(usage="OpenEO AsyncTask --task <task>",
                                      formatter_class=argparse.ArgumentDefaultsHelpFormatter)
@@ -49,13 +72,17 @@ def main():
     args = parser.parse_args()
 
     task = json.loads(args.task_json)
-    task_id = task["task_id"]
-    if task_id not in ['delete_batch_process_results']:
+    task_id = task['task_id']
+    if task_id not in [TASK_DELETE_BATCH_PROCESS_RESULTS, TASK_POLL_SENTINELHUB_BATCH_PROCESSES]:
         raise ValueError(f'unsupported task_id "{task_id}"')
 
-    arguments = task['arguments']
-    batch_job_id = arguments['batch_job_id']
-    subfolders = arguments['subfolders']
+    arguments = task.get('arguments', {})
+
+    # clunky but ConfigParams needs it
+    for envar, value in task.get('environment', {}).items():
+        os.environ[envar] = value
+
+    _log.info("ConfigParams(): {c}".format(c=ConfigParams()))
 
     java_opts = [
         "-client",
@@ -69,8 +96,28 @@ def main():
 
     batch_jobs = GpsBatchJobs(catalog=None, jvm=java_gateway.jvm, principal=None, key_tab=None)
 
-    _log.info(f"removing subfolders {subfolders} for batch job {batch_job_id}...")
-    batch_jobs.delete_batch_process_results(job_id=batch_job_id, subfolders=subfolders, propagate_errors=True)
+    if task_id == TASK_DELETE_BATCH_PROCESS_RESULTS:
+        batch_job_id = arguments['batch_job_id']
+        subfolders = arguments['subfolders']
+
+        _log.info(f"removing subfolders {subfolders} for batch job {batch_job_id}...")
+        batch_jobs.delete_batch_process_results(job_id=batch_job_id, subfolders=subfolders, propagate_errors=True)
+    elif task_id == TASK_POLL_SENTINELHUB_BATCH_PROCESSES:
+        batch_job_id = arguments['batch_job_id']
+        user_id = arguments['user_id']
+
+        while True:
+            time.sleep(SENTINEL_HUB_BATCH_PROCESSES_POLL_INTERVAL_S)
+
+            with JobRegistry() as registry:
+                job_info = registry.get_job(batch_job_id, user_id)
+
+            if job_info.get('dependency_status') not in ['awaiting', "awaiting_retry"]:
+                break
+            else:
+                batch_jobs.poll_sentinelhub_batch_processes(job_info)
+    else:
+        raise AssertionError(f'unexpected task_id "{task_id}"')
 
 
 if __name__ == '__main__':
